In addition to merging with Readline 8.1, this also enables colorization
of the input line based on argmatchers.  There are some incomplete parts
that hurt performance; they will be fixed very soon.

+++ readline/readline/examples/fileman.c
@@ -377,11 +377,11 @@ com_stat (arg)
 
   printf ("Statistics for `%s':\n", arg);
 
-  printf ("%s has %d link%s, and is %d byte%s in length.\n",
+  printf ("%s has %d link%s, and is %lu byte%s in length.\n",
 	  arg,
           finfo.st_nlink,
           (finfo.st_nlink == 1) ? "" : "s",
-          finfo.st_size,
+          (unsigned long)finfo.st_size,
           (finfo.st_size == 1) ? "" : "s");
   printf ("Inode Last Change at: %s", ctime (&finfo.st_ctime));
   printf ("      Last access at: %s", ctime (&finfo.st_atime));
